desactivate option

+++ gitlab/o2s-install.sh
@@ -4,10 +4,12 @@
 function afficher_aide() {
     echo "Utilisation :"
     echo "  ./o2s-install.sh [-o]"
+    echo
     echo "Installe omniauth2-shibboleth sur une instance Gitlab."
     echo
     echo "Options :"
-    echo "  -o : installe omniauth-shibboleth version 1.3.0"
+    echo "  -o : installe omniauth-shibboleth version 1.3.0."
+    echo "  -d : désactivation d'omniauth2-shibboleth."
     echo "  -h : affiche ce message d'aide." 
     exit 1
 }
@@ -39,16 +41,87 @@ insert_line() {
     fi
 }
 
-# Affichage de l'état d'un fichier
+# Fonction d'affichage de l'état d'un fichier
 display_state() {
     local file="$1"
     local state="$2"
 
     if [ "$state" -gt "0" ]; then
-        echo -e "${GREEN}$file mis à jour.${NOCOLOR}"
+        echo -e "${GREEN}$(basename $file) mis à jour.${NOCOLOR}"
     else
-        echo -e "${ORANGE}$file non modifié.${NOCOLOR}"
+        echo -e "${ORANGE}$(basename $file) non modifié.${NOCOLOR}"
     fi
+}
+
+# Fonction pour commenter les lignes correspondantes
+comment_lines() {
+    local filename="$1"
+    local search_string="$2"
+
+    rm -f temp
+    while IFS= read -r line; do
+        if [[ "$line" =~ "$search_string" ]] && [[ "$line" != \#* ]]; then
+            echo "#$line" >> temp
+        else
+            echo "$line" >> temp
+        fi
+    done < "$filename"
+    mv temp "$filename"
+}
+
+# Fonction pour décommenter les lignes correspondantes
+uncomment_lines() {
+    local filename="$1"
+    local search_string="$2"
+
+    awk -v search="$search_string" -v comment="#" '{ if ($0 ~ search) gsub("^" comment, "", $0); print }' "$filename" > temp
+    mv temp "$filename"
+}
+
+# Fonction de fin du script
+exit_script() {
+    /opt/gitlab/embedded/bin/bundle config unset frozen
+    echo "Reconfigure Gitlab..."
+    gitlab-ctl reconfigure > /dev/null 2>&1
+    gitlab-ctl restart puma
+    gitlab-ctl restart sidekiq
+
+    # URL à vérifier
+    URL=$(echo $(cat /etc/gitlab/gitlab.rb | grep "^external_url") | cut -d'"' -f2)
+    # Durée maximale en secondes
+    max_duration=180
+
+    # Début du compte à rebours
+    start_time=$(date +%s)
+
+    # Exécution de la commande "curl" pour récupérer le code HTTP
+    http_code=0
+    while [ "$http_code" -ne 200 ]; do
+        # Calcul de la durée écoulée
+        current_time=$(date +%s)
+        duration=$((current_time - start_time))
+
+        # Vérification si la durée maximale est dépassée
+        if [ "$duration" -gt "$max_duration" ]; then
+            echo -e "${RED}Timeout: la durée maximale de $max_duration secondes est dépassée.${NOCOLOR}"
+            gitlab-ctl status | grep "^down"
+            exit 1
+        fi
+
+        # Récupération du code HTTP
+        sleep 10
+        http_code=$(curl -sL -w "%{http_code}\\n" "$URL" -o /dev/null)
+
+        # Arrêt de la temporisation lorsque le code HTTP est 200
+        if [ "$http_code" -eq 200 ]; then
+            echo -e "\r${GREEN}$URL : $http_code${NOCOLOR}"
+            break
+        else
+            echo -ne "\r${RED}$URL : $http_code${NOCOLOR}"
+        fi
+    done
+
+    exit 0
 }
 
 # Colorisation des sorties
@@ -59,16 +132,23 @@ ORANGE='\033[0;33m'
 
 # Paramètres du script
 DIR="/opt/gitlab/embedded/service/gitlab-rails"
+CONF="/etc/gitlab/gitlab.rb"
 OMNIAUTH_VERSION="2.1.0"
 
 # Traitement des options de ligne de commande
-while getopts "oh" opt; do
+while getopts "odh" opt; do
     case $opt in
         o)
             OMNIAUTH_SHIBBOLETH_VERSION="1.3.0"
             OMNIAUTH_SHIBBOLETH_CHECKSUM="b0bb725ced5cb76fbfc187ddbb8ad6864d0cd5df714cab36a528df8ee4b1d113"
             OLD_GEM="omniauth2-shibboleth"
             NEW_GEM="omniauth-shibboleth"
+            ;;
+        d)  
+            comment_lines "$CONF" "o2s_comment"
+            i=$(delete_line "$DIR/app/helpers/auth_helper.rb" "    shibboleth")
+            display_state "$DIR/app/helpers/auth_helper.rb" "$i"
+            exit_script
             ;;
         h)
             afficher_aide
@@ -97,10 +177,14 @@ then
 fi
 
 # Copie du logo Shibboleth
-mkdir -p $DIR/public/images/auth_buttons/
-cp ./shibboleth_64.png $DIR/public/images/auth_buttons/
+if [ -f ./shibboleth_64.png ]; then
+    mkdir -p $DIR/public/images/auth_buttons/
+    cp ./shibboleth_64.png $DIR/public/images/auth_buttons/
+fi
 
 # Modification des fichiers de configuration de Gitlab
+uncomment_lines "$CONF" "o2s_comment"
+
 i=$(insert_line "$DIR/app/helpers/auth_helper.rb" "    shibboleth" "    salesforce")
 display_state "$DIR/app/helpers/auth_helper.rb" "$i"
 
@@ -118,45 +202,4 @@ k=$(insert_line "$DIR/Gemfile.lock" "      omniauth (>= $OMNIAUTH_VERSION)" "   
 l=$(insert_line "$DIR/Gemfile.lock" "    $NEW_GEM ($OMNIAUTH_SHIBBOLETH_VERSION)" "      ruby-saml *")
 display_state "$DIR/Gemfile.lock" "$((i+j+k+l))"
 
-# Redemarrage
-/opt/gitlab/embedded/bin/bundle config unset frozen
-gitlab-ctl stop
-gitlab-ctl start &
-
-# URL à vérifier
-URL=$(echo $(cat /etc/gitlab/gitlab.rb | grep "^external_url") | cut -d'"' -f2)
-# Durée maximale en secondes
-max_duration=180
-
-# Début du compte à rebours
-start_time=$(date +%s)
-
-# Exécution de la commande "curl" pour récupérer le code HTTP
-http_code=0
-while [ "$http_code" -ne 200 ]; do
-    # Calcul de la durée écoulée
-    current_time=$(date +%s)
-    duration=$((current_time - start_time))
-
-    # Vérification si la durée maximale est dépassée
-    if [ "$duration" -gt "$max_duration" ]; then
-        echo -e "${RED}Timeout: la durée maximale de $max_duration secondes est dépassée.${NOCOLOR}"
-        exit 1
-    fi
-
-    # Récupération du code HTTP
-    sleep 10
-    http_code=$(curl -sL -w "%{http_code}\\n" "$URL" -o /dev/null)
-
-    # Arrêt de la temporisation lorsque le code HTTP est 200
-    if [ "$http_code" -eq 200 ]; then
-        echo -e "\r${GREEN}$URL : $http_code${NOCOLOR}"
-        break
-    else
-        echo -ne "\r${RED}$URL : $http_code${NOCOLOR}"
-    fi
-done
-
-# Sortie
-echo "$(/opt/gitlab/embedded/bin/gem list | grep shibboleth) installé"
-exit 0
+exit_script
